fixed structure issue
