--- to-do-list/src/main/java/com/example/to_do_list/ToDoListApplication.java
@@ -1,13 +1,0 @@
-package com.example.to_do_list;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ToDoListApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ToDoListApplication.class, args);
-	}
-
-}
