--- to-do-list/src/main/java/com/example/to_do_list/controllers/AuthController.java
@@ -1,47 +1,0 @@
-package com.example.to_do_list.controllers;
-
-
-import com.example.to_do_list.dto.AuthResponseDTO;
-import com.example.to_do_list.dto.LoginDTO;
-import com.example.to_do_list.dto.RegisterDTO;
-import com.example.to_do_list.service.AuthService;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.web.bind.annotation.*;
-
-import java.io.IOException;
-
-@RestController
-@RequestMapping("/auth")
-public class AuthController {
-
-    private AuthService authService;
-
-    public AuthController(AuthService authService) {
-        this.authService = authService;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@RequestBody LoginDTO loginDto) {
-        return authService.authenticateUser(loginDto);
-    }
-
-    @PostMapping("/register")
-    public ResponseEntity<?> register(@RequestBody RegisterDTO registerDto) {
-        return authService.registerUser(registerDto);
-    }
-
-    @PostMapping("/logout")
-    public ResponseEntity<?> logout(@RequestHeader("Authorization") String token) {
-        return authService.logout(token);
-    }
-
-    @GetMapping("/google")
-    public void redirectToGoogleAuth(HttpServletResponse response) throws IOException {
-        response.sendRedirect("/oauth2/authorization/google");
-    }
-
-}
