--- to-do-list/src/main/java/com/example/to_do_list/dto/RegisterDTO.java
@@ -1,25 +1,0 @@
-package com.example.to_do_list.dto;
-
-import lombok.Data;
-
-
-public class RegisterDTO {
-    private String username;
-    private String password;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
