--- to-do-list/src/main/java/com/example/to_do_list/models/Status.java
@@ -1,7 +1,0 @@
-package com.example.to_do_list.models;
-
-public enum Status {
-    TODO,
-    IN_PROGRESS,
-    COMPLETED
-}
