--- to-do-list/src/main/java/com/example/to_do_list/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.example.to_do_list.repository;
-
-
-import com.example.to_do_list.models.UserEntity;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends MongoRepository<UserEntity, String> {
-    Optional<UserEntity> findByUsername(String username);
-
-    Optional<UserEntity> findByEmail(String email);
-
-    Boolean existsByUsername(String username);
-
-}
