--- to-do-list/src/main/java/com/example/to_do_list/security/CustomUserDetailsService.java
@@ -1,31 +1,0 @@
-package com.example.to_do_list.security;
-
-import com.example.to_do_list.models.UserEntity;
-import com.example.to_do_list.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private UserRepository userRepository;
-
-    @Autowired
-    public CustomUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        UserEntity user = userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("Username not found"));
-        return new User(user.getUsername(), user.getPassword(), new ArrayList<>());
-    }
-
-}
