--- to-do-list/src/main/java/com/example/to_do_list/security/SecurityConstants.java
@@ -1,8 +1,0 @@
-package com.example.to_do_list.security;
-
-import io.github.cdimascio.dotenv.Dotenv;
-
-public class SecurityConstants {
-    public static final long JWT_EXPIRATION = 864000000;
-    public static final String SECRET_KEY = "5f1b41f01a99eef508a94bff09bedbe975e44abcf6f4cfbf73307b4ab5895328";
-}
