+++ to-do-list/src/main/java/com/masteryhub/todoapp/ToDoListApplication.java
@@ -1,0 +1,13 @@
+package com.masteryhub.todoapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ToDoListApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ToDoListApplication.class, args);
+	}
+
+}
