+++ to-do-list/src/main/java/com/masteryhub/todoapp/config/SwaggerConfig.java
@@ -1,0 +1,36 @@
+package com.masteryhub.todoapp.config;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+@Configuration
+@EnableWebMvc
+public class SwaggerConfig implements WebMvcConfigurer {
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2).select()
+                .apis(RequestHandlerSelectors.basePackage("com.example.to_do_list.controller"))
+                .paths(PathSelectors.regex("/.*"))
+                .build().apiInfo(apiInfoMetaData());
+    }
+
+    private ApiInfo apiInfoMetaData() {
+
+        return new ApiInfoBuilder().title("TODO List API")
+                .description("API Endpoint Decoration")
+                .contact(new Contact("Dev-Team", "https://www.masteryhub-its.com/", "abdulrahmantarek0.1@gmail.com"))
+                .license("Apache 2.0")
+                .licenseUrl("http://www.apache.org/licenses/LICENSE-2.0.html")
+                .version("1.0.0")
+                .build();
+    }
+}
