+++ to-do-list/src/main/java/com/masteryhub/todoapp/dto/AuthResponseDTO.java
@@ -1,0 +1,26 @@
+package com.masteryhub.todoapp.dto;
+
+public class AuthResponseDTO {
+    private String accessToken;
+    private String tokenType = "Bearer ";
+
+    public AuthResponseDTO(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+
+    public void setAccessToken(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getTokenType() {
+        return tokenType;
+    }
+
+    public void setTokenType(String tokenType) {
+        this.tokenType = tokenType;
+    }
+}
