+++ to-do-list/src/main/java/com/masteryhub/todoapp/dto/LoginDTO.java
@@ -1,0 +1,25 @@
+package com.masteryhub.todoapp.dto;
+
+import lombok.Data;
+
+
+public class LoginDTO {
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
