+++ to-do-list/src/main/java/com/masteryhub/todoapp/models/Status.java
@@ -1,0 +1,7 @@
+package com.masteryhub.todoapp.models;
+
+public enum Status {
+    TODO,
+    IN_PROGRESS,
+    COMPLETED
+}
