+++ to-do-list/src/main/java/com/masteryhub/todoapp/models/TodoEntity.java
@@ -1,0 +1,45 @@
+package com.masteryhub.todoapp.models;
+
+
+public class TodoEntity {
+
+    private Long id;
+
+    private String title;
+
+    private String description;
+
+    private Status status;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Status getStatus() {
+        return status;
+    }
+
+    public void setStatus(Status status) {
+        this.status = status;
+    }
+}
