+++ to-do-list/src/main/java/com/masteryhub/todoapp/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package com.masteryhub.todoapp.repository;
+
+
+import com.masteryhub.todoapp.models.UserEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends MongoRepository<UserEntity, String> {
+    Optional<UserEntity> findByUsername(String username);
+
+    Optional<UserEntity> findByEmail(String email);
+
+    Boolean existsByUsername(String username);
+
+}
