+++ to-do-list/src/main/java/com/masteryhub/todoapp/security/SecurityConstants.java
@@ -1,0 +1,8 @@
+package com.masteryhub.todoapp.security;
+
+import io.github.cdimascio.dotenv.Dotenv;
+
+public class SecurityConstants {
+    public static final long JWT_EXPIRATION = 864000000;
+    public static final String SECRET_KEY = "5f1b41f01a99eef508a94bff09bedbe975e44abcf6f4cfbf73307b4ab5895328";
+}
