--- to-do-list/src/test/java/com/example/to_do_list/ToDoListApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.to_do_list;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ToDoListApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
