+++ to-do-list/src/test/java/com/masteryhub/todoapp/ToDoListApplicationTests.java
@@ -1,0 +1,13 @@
+package com.masteryhub.todoapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ToDoListApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
